+++ hydra-node/src/Hydra/Ledger/Cardano/Evaluate.hs
@@ -18,36 +18,31 @@ import qualified Cardano.Api.UTxO as UTxO
 import qualified Cardano.Ledger.Alonzo.Data as Ledger
 import Cardano.Ledger.Alonzo.Language (Language (PlutusV1, PlutusV2))
 import qualified Cardano.Ledger.Alonzo.PlutusScriptApi as Ledger
-import Cardano.Ledger.Alonzo.Scripts (CostModels (CostModels), ExUnits (..), Prices (..), txscriptfee)
+import Cardano.Ledger.Alonzo.Scripts (CostModels (CostModels), txscriptfee)
 import Cardano.Ledger.Alonzo.TxInfo (slotToPOSIXTime)
-import qualified Cardano.Ledger.Alonzo.TxInfo as Ledger
-import Cardano.Ledger.Babbage.PParams (PParams' (..))
-import qualified Cardano.Ledger.Babbage.Tx as Babbage
-import Cardano.Ledger.BaseTypes (ProtVer (..), boundRational)
 import Cardano.Ledger.Coin (Coin (Coin))
 import Cardano.Ledger.Val (Val ((<+>)), (<×>))
 import Cardano.Slotting.EpochInfo (EpochInfo, fixedEpochInfo)
 import Cardano.Slotting.Slot (EpochNo (EpochNo), EpochSize (EpochSize), SlotNo (SlotNo))
 import Cardano.Slotting.Time (RelativeTime (RelativeTime), SlotLength (getSlotLength), SystemStart (SystemStart), mkSlotLength, toRelativeTime)
-import Control.Arrow (left)
 import qualified Data.ByteString as BS
 import Data.Default (def)
 import qualified Data.Map as Map
-import Data.Maybe (fromJust)
 import Data.Ratio ((%))
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
 import Flat (flat)
 import Hydra.Cardano.Api (
+  CardanoEra (BabbageEra),
   CardanoMode,
   ConsensusMode (CardanoMode),
   Era,
   EraHistory (EraHistory),
-  EraInMode (BabbageEraInCardanoMode),
+  ExecutionUnitPrices (..),
   ExecutionUnits (..),
   IsShelleyBasedEra (shelleyBasedEra),
-  LedgerEra,
+  LedgerEpochInfo (..),
   Lovelace,
-  ProtocolParameters (protocolParamMaxTxExUnits, protocolParamMaxTxSize),
+  ProtocolParameters (..),
   ScriptExecutionError (ScriptErrorMissingScript),
   ScriptWitnessIndex,
   SerialiseAsCBOR (serialiseToCBOR),
@@ -55,11 +50,14 @@ import Hydra.Cardano.Api (
   TransactionValidityError,
   Tx,
   UTxO,
+  bundleProtocolParams,
   evaluateTransactionExecutionUnits,
+  fromAlonzoCostModels,
   fromLedgerCoin,
   fromLedgerPParams,
   getTxBody,
   shelleyBasedEra,
+  toAlonzoPrices,
   toLedgerEpochInfo,
   toLedgerExUnits,
   toLedgerPParams,
@@ -80,13 +78,16 @@ import Ouroboros.Consensus.HardFork.History (
   mkInterpreter,
  )
 import Ouroboros.Consensus.Util.Counting (NonEmpty (NonEmptyOne))
-import Plutus.ApiCommon (mkTermToEvaluate)
-import qualified Plutus.ApiCommon as Plutus
 import qualified PlutusCore as PLC
+import qualified PlutusLedgerApi.Common as Plutus
 import Test.Cardano.Ledger.Alonzo.PlutusScripts (testingCostModelV1, testingCostModelV2)
 import Test.QuickCheck (choose)
 import Test.QuickCheck.Gen (chooseWord64)
 import qualified UntypedPlutusCore as UPLC
+
+-- Global import for HasField instances, which seem to not come into scope
+-- otherwise. Used by collectTwoPhaseScriptInputs in prepareTxScripts
+import Cardano.Ledger.Babbage.PParams
 
 -- * Evaluate transactions
 
@@ -119,10 +120,9 @@ evaluateTx' maxUnits tx utxo =
  where
   result =
     evaluateTransactionExecutionUnits
-      BabbageEraInCardanoMode
       systemStart
-      eraHistory
-      pparams{protocolParamMaxTxExUnits = Just maxUnits}
+      (toLedgerEpochInfo eraHistory)
+      (bundleProtocolParams BabbageEra pparams{protocolParamMaxTxExUnits = Just maxUnits})
       (UTxO.toApi utxo)
       (getTxBody tx)
 
@@ -184,9 +184,9 @@ usedExecutionUnits report =
   budgets = rights $ toList report
 
 -- | Estimate minimum fee for given transaction and evaluated redeemers. Instead
--- of using the budgets from the transaction (which might are usually set to 0
--- until balancing), this directly computes the fee from transaction size and
--- the units of the 'EvaluationReport'. Note that this function only provides a
+-- of using the budgets from the transaction (which are usually set to 0 until
+-- balancing), this directly computes the fee from transaction size and the
+-- units of the 'EvaluationReport'. Note that this function only provides a
 -- rough estimate using this modules' 'pparams' and likely under-estimates cost
 -- as we have no witnesses on this 'Tx'.
 estimateMinFee ::
@@ -196,12 +196,14 @@ estimateMinFee ::
 estimateMinFee tx evaluationReport =
   fromLedgerCoin $
     (txSize <×> a <+> b)
-      <+> txscriptfee (_prices pp) allExunits
+      <+> txscriptfee prices allExunits
  where
   txSize = BS.length $ serialiseToCBOR tx
-  a = Coin . fromIntegral $ _minfeeA pp
-  b = Coin . fromIntegral $ _minfeeB pp
-  pp = toLedgerPParams (shelleyBasedEra @Era) pparams
+  a = Coin . fromIntegral $ protocolParamTxFeePerByte pparams
+  b = Coin . fromIntegral $ protocolParamTxFeeFixed pparams
+  prices =
+    fromMaybe (error "no prices in protocol param fixture") $
+      toAlonzoPrices =<< protocolParamPrices pparams
   allExunits = foldMap toLedgerExUnits . rights $ toList evaluationReport
 
 -- * Profile transactions
@@ -212,6 +214,7 @@ estimateMinFee tx evaluationReport =
 -- to gather profiling information.
 --
 -- NOTE: This assumes we use 'Babbage' and only 'PlutusV2' scripts are used.
+{-# DEPRECATED prepareTxScripts "This function is currently broken due to missing dependencies" #-}
 prepareTxScripts ::
   Tx ->
   UTxO ->
@@ -225,53 +228,77 @@ prepareTxScripts tx utxo = do
 
   -- Fully applied UPLC programs which we could run using the cekMachine
   programs <- forM results $ \(script, _language, arguments, _exUnits, _costModel) -> do
-    let pv = Ledger.transProtocolVersion (_protocolVersion pp)
-        pArgs = Ledger.getPlutusData <$> arguments
-    appliedTerm <- left show $ mkTermToEvaluate Plutus.PlutusV2 pv script pArgs
+    let pArgs = Ledger.getPlutusData <$> arguments
+    appliedTerm <- mkTermToEvaluate Plutus.PlutusV2 protocolVersion script pArgs
     pure $ UPLC.Program () (PLC.defaultVersion ()) appliedTerm
 
   pure $ flat <$> programs
  where
   pp = toLedgerPParams (shelleyBasedEra @Era) pparams
-  ltx = toLedgerTx tx :: Babbage.ValidatedTx LedgerEra
+
+  ltx = toLedgerTx tx
+
   lutxo = toLedgerUTxO utxo
-  ei = toLedgerEpochInfo eraHistory
+
+  LedgerEpochInfo ei = toLedgerEpochInfo eraHistory
+
+  protocolVersion =
+    let (major, minor) = protocolParamProtocolVersion pparams
+     in Plutus.ProtocolVersion
+          { Plutus.pvMajor = fromIntegral major
+          , Plutus.pvMinor = fromIntegral minor
+          }
+
+  mkTermToEvaluate :: a -> b -> c -> d -> m (UPLC.Term UPLC.NamedDeBruijn UPLC.DefaultUni UPLC.DefaultFun ())
+  mkTermToEvaluate =
+    error "mkTermToEvaluate not yet exported, need to wait for plutus-ledger-api 1.2.0.0 support in cardano-api"
 
 -- * Fixtures
 
--- | Current mainchain protocol parameters.
+-- | Current (2023-04-12) mainchain protocol parameters.
+-- XXX: Avoid specifiying not required parameters here (e.g. max block units
+-- should not matter).
 pparams :: ProtocolParameters
 pparams =
-  fromLedgerPParams (shelleyBasedEra @Era) $
-    def
-      { _costmdls =
-          CostModels $
-            Map.fromList
-              [ (PlutusV1, testingCostModelV1)
-              , (PlutusV2, testingCostModelV2)
-              ]
-      , _maxValSize = 1000000000
-      , _maxTxExUnits = ExUnits 14_000_000 10_000_000_000
-      , _maxBlockExUnits = ExUnits 62_000_000 40_000_000_000
-      , _protocolVersion = ProtVer 7 0
-      , _maxTxSize = 16384
-      , _minfeeA = 44
-      , _minfeeB = 155381
-      , _prices =
-          Prices
-            { prSteps = fromJust $ boundRational $ 721 % 10000000
-            , prMem = fromJust $ boundRational $ 577 % 10000
+  (fromLedgerPParams (shelleyBasedEra @Era) def)
+    { protocolParamCostModels =
+        fromAlonzoCostModels
+          . CostModels
+          $ Map.fromList
+            [ (PlutusV1, testingCostModelV1)
+            , (PlutusV2, testingCostModelV2)
+            ]
+    , protocolParamMaxTxExUnits = Just maxTxExecutionUnits
+    , protocolParamMaxBlockExUnits =
+        Just
+          ExecutionUnits
+            { executionMemory = 62_000_000
+            , executionSteps = 40_000_000_000
             }
-      }
+    , protocolParamProtocolVersion = (7, 0)
+    , protocolParamMaxTxSize = maxTxSize
+    , protocolParamMaxValueSize = Just 1000000000
+    , protocolParamTxFeePerByte = 44 -- a
+    , protocolParamTxFeeFixed = 155381 -- b
+    , protocolParamPrices =
+        Just
+          ExecutionUnitPrices
+            { priceExecutionSteps = 721 % 10000000
+            , priceExecutionMemory = 577 % 10000
+            }
+    }
 
 -- | Max transaction size of the current 'pparams'.
 maxTxSize :: Natural
-maxTxSize = protocolParamMaxTxSize pparams
+maxTxSize = 16384
 
 -- | Max transaction execution unit budget of the current 'params'.
 maxTxExecutionUnits :: ExecutionUnits
 maxTxExecutionUnits =
-  fromJust $ protocolParamMaxTxExUnits pparams
+  ExecutionUnits
+    { executionMemory = 14_000_000
+    , executionSteps = 10_000_000_000
+    }
 
 -- | Max memory and cpu units of the current 'pparams'.
 maxMem, maxCpu :: Natural
@@ -280,6 +307,7 @@ maxMem = executionMemory maxTxExecutionUnits
 
 -- | An artifical era history comprised by a single never ending (forking) era,
 -- with fixed 'epochSize' and 'slotLength'.
+-- TODO: can we remove 'EraHistory' fixtures in exchange for 'LedgerEpochInfo' versions?
 eraHistory :: EraHistory CardanoMode
 eraHistory =
   EraHistory CardanoMode (mkInterpreter summary)